Had to refactor due to issues with the this keyword, I am not sure
what I was originally thinking. I didn't want to have to pass the graph
to every function call. The issue has been resolved by using one
interface where the graph is a property of the interface. This allows
all of our methods to be purer functions and not have to worry about
passing the graph around.

The interface has all of the methods under it rather that having the
graph, vertex, edge, and walk interfaces. This doesn't make our
code as modular as I would like, but it does make it easier to use.

The Euler path and Euler circuit functions have also been implemented.
We can determine if a provided sequence of vertices is a valid Euler
path or circuit. We can also find an Euler path or circuit if one
exists and return it as a sequence of vertex ids.

The Hamiltonian functions need to be implemented but I wanted to
get a commit in.