Addressed an issue with the hamiltonian path generator.
Added the ability to generate a weighted graph.
We are working on trees, tree traversals, and spanning trees.
Created a utility file for generic non-graph related functions.

Still working on trees but wanted to save this progress. These topics
were not covered last week or I would have included them from the start.