Graph manipulation functionality has been completed. A shortest
path algorithm has been implemented. The shortest path algorithm uses
dijsktra's algorithm.

Previous functionality was adjusted here and there. Fuzzy testing needs
to be implemented so we can ensure the code doesn't break when
bad inputs are given, but all tests as written currently pass.

More tests are needed, and the library, now that I have a better of how
it should all work together, needs to be refactored. It will be
be refactored to use Classes to model the necessary data structures and
the basic ability to create a graph, add nodes, add edges, and remove
nodes and edges will be implemented.

All the graph utility or analysis functions will be implemented in a
functional manner. I think this will be a better approach for the
library.